+++ game/src/player.lua
@@ -5,8 +5,8 @@ function Player:new()
     self.speed = 500
     self.width = self.image:getWidth()
     self.height = self.image:getHeight();
-    self.x = (window_width - self.width) / 2 
-    self.y = window_height - (self.height * 2)
+    self.x = (WINDOW_WIDTH - self.width) / 2
+    self.y = WINDOW_HEIGHT - self.height * 2
 end
 
 function Player:update(dt)
@@ -29,14 +29,14 @@ end
 function Player:limit()
     if self.x < 0 then
         self.x = 0
-    elseif self.x + self.width > window_width then
-        self.x = window_width - self.width
+    elseif self.x + self.width > WINDOW_WIDTH then
+        self.x = WINDOW_WIDTH - self.width
     end
 
     if self.y < 0 then
         self.y = 0
-    elseif self.y + self.height > window_height then
-        self.y = window_height - self.height
+    elseif self.y + self.height > WINDOW_HEIGHT then
+        self.y = WINDOW_HEIGHT - self.height
     end
 end
 
@@ -46,19 +46,19 @@ end
 
 function Player:collided(enemy)
     -- allow small amounts of grace
-    local e_left = enemy.x + 1
-    local e_right = enemy.x + enemy.width - 2
-    local e_top = enemy.y + 1
-    local e_bottom = enemy.y + enemy.height - 2
+    local eLeft = enemy.x + 1
+    local eRight = enemy.x + enemy.width - 2
+    local eTop = enemy.y + 1
+    local eBottom = enemy.y + enemy.height - 2
 
     -- only consider the main trunk of the ship
-    local p_left = self.x + 9
-    local p_right = self.x + self.width - 9
-    local p_top = self.y + 13
-    local p_bottom = self.y + self.height - 13
+    local pLeft = self.x + 9
+    local pRight = self.x + self.width - 9
+    local pTop = self.y + 13
+    local pBottom = self.y + self.height - 13
 
-    return  e_right > p_left
-        and e_left < p_right
-        and e_bottom > p_top
-        and e_top < p_bottom
+    return  eRight > pLeft
+        and eLeft < pRight
+        and eBottom > pTop
+        and eTop < pBottom
 end